connection-manager: removed assertions in demotedToColdRemote
'bidirectional_IO' test case tripped over them.  I don't think that this
is related to the changes in this PR.  They probably only surfaced now
because of the changes in 71ca24c7a.  The connection manager and
inbound-governor can deal with the situation when these assertion fail,
so they are not really needed.